Improve SequentialEvent tests
I am still not sure how to properly test the "factory" methods of
events, such as SequentialEvent.after or 'or' or 'failIf'. Need we only
test a new event was instantiated with a certain configuration? Or
should we completely black box the internals and assert the expected
behavior of those new events returned? But then this gets into testing
things that are already being tested in _those other classes'_ unit
tests.